moved computeNodesConfigFromList to nodes coordinator lite
