+++ sharding/indexHashedNodesCoordinator_test.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"runtime"
-	"sort"
 	"strconv"
 	"sync"
 	"sync/atomic"
@@ -1937,381 +1936,6 @@ func TestIndexHashedNodesCoordinator_ShuffleOutNilConfig(t *testing.T) {
 	require.False(t, processCalled)
 	expectedShardForNotFound := uint32(0)
 	require.Equal(t, expectedShardForNotFound, newShard)
-}
-
-func TestIndexHashedNodesCoordinator_computeNodesConfigFromListNilPreviousNodesConfig(t *testing.T) {
-	t.Parallel()
-
-	arguments := createArguments()
-	pk := []byte("pk")
-	arguments.SelfPublicKey = pk
-	ihgs, _ := NewIndexHashedNodesCoordinator(arguments)
-
-	ihgs.flagWaitingListFix.Unset()
-	validatorInfos := make([]*state.ShardValidatorInfo, 0)
-	newNodesConfig, err := ihgs.computeNodesConfigFromList(nil, validatorInfos)
-
-	assert.Nil(t, newNodesConfig)
-	assert.False(t, errors.Is(err, ErrNilPreviousEpochConfig))
-
-	newNodesConfig, err = ihgs.computeNodesConfigFromList(nil, nil)
-
-	assert.Nil(t, newNodesConfig)
-	assert.False(t, errors.Is(err, ErrNilPreviousEpochConfig))
-
-	ihgs.flagWaitingListFix.Set()
-	newNodesConfig, err = ihgs.computeNodesConfigFromList(nil, validatorInfos)
-
-	assert.Nil(t, newNodesConfig)
-	assert.True(t, errors.Is(err, ErrNilPreviousEpochConfig))
-
-	newNodesConfig, err = ihgs.computeNodesConfigFromList(nil, nil)
-
-	assert.Nil(t, newNodesConfig)
-	assert.True(t, errors.Is(err, ErrNilPreviousEpochConfig))
-}
-
-func TestIndexHashedNodesCoordinator_computeNodesConfigFromListNoValidators(t *testing.T) {
-	t.Parallel()
-
-	arguments := createArguments()
-	pk := []byte("pk")
-	arguments.SelfPublicKey = pk
-	ihgs, _ := NewIndexHashedNodesCoordinator(arguments)
-
-	validatorInfos := make([]*state.ShardValidatorInfo, 0)
-	newNodesConfig, err := ihgs.computeNodesConfigFromList(&nodesCoordinator.EpochNodesConfig{}, validatorInfos)
-
-	assert.Nil(t, newNodesConfig)
-	assert.True(t, errors.Is(err, ErrMapSizeZero))
-
-	newNodesConfig, err = ihgs.computeNodesConfigFromList(&nodesCoordinator.EpochNodesConfig{}, nil)
-
-	assert.Nil(t, newNodesConfig)
-	assert.True(t, errors.Is(err, ErrMapSizeZero))
-}
-
-func TestIndexHashedNodesCoordinator_computeNodesConfigFromListNilPk(t *testing.T) {
-	t.Parallel()
-
-	arguments := createArguments()
-	pk := []byte("pk")
-	arguments.SelfPublicKey = pk
-	ihgs, _ := NewIndexHashedNodesCoordinator(arguments)
-
-	validatorInfos :=
-		[]*state.ShardValidatorInfo{
-			{
-				PublicKey:  pk,
-				ShardId:    0,
-				List:       "test1",
-				Index:      0,
-				TempRating: 0,
-			},
-			{
-				PublicKey:  nil,
-				ShardId:    0,
-				List:       "test",
-				Index:      0,
-				TempRating: 0,
-			},
-		}
-
-	newNodesConfig, err := ihgs.computeNodesConfigFromList(&nodesCoordinator.EpochNodesConfig{}, validatorInfos)
-
-	assert.Nil(t, newNodesConfig)
-	assert.NotNil(t, err)
-	assert.Equal(t, ErrNilPubKey, err)
-}
-
-func TestIndexHashedNodesCoordinator_computeNodesConfigFromListValidatorsWithFix(t *testing.T) {
-	t.Parallel()
-
-	arguments := createArguments()
-	pk := []byte("pk")
-	arguments.SelfPublicKey = pk
-	ihgs, _ := NewIndexHashedNodesCoordinator(arguments)
-	ihgs.flagWaitingListFix.Set()
-
-	shard0Eligible0 := &state.ShardValidatorInfo{
-		PublicKey:  []byte("pk0"),
-		List:       string(common.EligibleList),
-		Index:      1,
-		TempRating: 2,
-		ShardId:    0,
-	}
-	shard0Eligible1 := &state.ShardValidatorInfo{
-		PublicKey:  []byte("pk1"),
-		List:       string(common.EligibleList),
-		Index:      2,
-		TempRating: 2,
-		ShardId:    0,
-	}
-	shardmetaEligible0 := &state.ShardValidatorInfo{
-		PublicKey:  []byte("pk2"),
-		ShardId:    core.MetachainShardId,
-		List:       string(common.EligibleList),
-		Index:      1,
-		TempRating: 4,
-	}
-	shard0Waiting0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk3"),
-		List:      string(common.WaitingList),
-		Index:     14,
-		ShardId:   0,
-	}
-	shardmetaWaiting0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk4"),
-		ShardId:   core.MetachainShardId,
-		List:      string(common.WaitingList),
-		Index:     15,
-	}
-	shard0New0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk5"),
-		List:      string(common.NewList), Index: 3,
-		ShardId: 0,
-	}
-	shard0Leaving0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk6"),
-		List:      string(common.LeavingList),
-		ShardId:   0,
-	}
-	shardMetaLeaving1 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk7"),
-		List:      string(common.LeavingList),
-		Index:     1,
-		ShardId:   core.MetachainShardId,
-	}
-
-	validatorInfos :=
-		[]*state.ShardValidatorInfo{
-			shard0Eligible0,
-			shard0Eligible1,
-			shardmetaEligible0,
-			shard0Waiting0,
-			shardmetaWaiting0,
-			shard0New0,
-			shard0Leaving0,
-			shardMetaLeaving1,
-		}
-
-	previousConfig := &nodesCoordinator.EpochNodesConfig{
-		EligibleMap: map[uint32][]validator{
-			0: {
-				mock.NewValidatorMock(shard0Eligible0.PublicKey, 0, 0),
-				mock.NewValidatorMock(shard0Eligible1.PublicKey, 0, 0),
-				mock.NewValidatorMock(shard0Leaving0.PublicKey, 0, 0),
-			},
-			core.MetachainShardId: {
-				mock.NewValidatorMock(shardmetaEligible0.PublicKey, 0, 0),
-			},
-		},
-		WaitingMap: map[uint32][]validator{
-			0: {
-				mock.NewValidatorMock(shard0Waiting0.PublicKey, 0, 0),
-			},
-			core.MetachainShardId: {
-				mock.NewValidatorMock(shardmetaWaiting0.PublicKey, 0, 0),
-				mock.NewValidatorMock(shardMetaLeaving1.PublicKey, 0, 0),
-			},
-		},
-	}
-
-	newNodesConfig, err := ihgs.computeNodesConfigFromList(previousConfig, validatorInfos)
-	assert.Nil(t, err)
-
-	assert.Equal(t, uint32(1), newNodesConfig.NbShards)
-
-	verifySizes(t, newNodesConfig)
-	verifyLeavingNodesInEligibleOrWaiting(t, newNodesConfig)
-
-	// maps have the correct validators inside
-	eligibleListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0Eligible0, shard0Eligible1, shard0Leaving0})
-	assert.Equal(t, eligibleListShardZero, newNodesConfig.EligibleMap[0])
-	eligibleListMeta := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shardmetaEligible0})
-	assert.Equal(t, eligibleListMeta, newNodesConfig.EligibleMap[core.MetachainShardId])
-
-	waitingListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0Waiting0})
-	assert.Equal(t, waitingListShardZero, newNodesConfig.WaitingMap[0])
-	waitingListMeta := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shardmetaWaiting0, shardMetaLeaving1})
-	assert.Equal(t, waitingListMeta, newNodesConfig.WaitingMap[core.MetachainShardId])
-
-	leavingListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0Leaving0})
-	assert.Equal(t, leavingListShardZero, newNodesConfig.LeavingMap[0])
-
-	leavingListMeta := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shardMetaLeaving1})
-	assert.Equal(t, leavingListMeta, newNodesConfig.LeavingMap[core.MetachainShardId])
-
-	newListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0New0})
-	assert.Equal(t, newListShardZero, newNodesConfig.NewList)
-}
-
-func TestIndexHashedNodesCoordinator_computeNodesConfigFromListValidatorsNoFix(t *testing.T) {
-	t.Parallel()
-
-	arguments := createArguments()
-	pk := []byte("pk")
-	arguments.SelfPublicKey = pk
-	ihgs, _ := NewIndexHashedNodesCoordinator(arguments)
-
-	shard0Eligible0 := &state.ShardValidatorInfo{
-		PublicKey:  []byte("pk0"),
-		List:       string(common.EligibleList),
-		Index:      1,
-		TempRating: 2,
-		ShardId:    0,
-	}
-	shard0Eligible1 := &state.ShardValidatorInfo{
-		PublicKey:  []byte("pk1"),
-		List:       string(common.EligibleList),
-		Index:      2,
-		TempRating: 2,
-		ShardId:    0,
-	}
-	shardmetaEligible0 := &state.ShardValidatorInfo{
-		PublicKey:  []byte("pk2"),
-		ShardId:    core.MetachainShardId,
-		List:       string(common.EligibleList),
-		Index:      1,
-		TempRating: 4,
-	}
-	shard0Waiting0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk3"),
-		List:      string(common.WaitingList),
-		Index:     14,
-		ShardId:   0,
-	}
-	shardmetaWaiting0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk4"),
-		ShardId:   core.MetachainShardId,
-		List:      string(common.WaitingList),
-		Index:     15,
-	}
-	shard0New0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk5"),
-		List:      string(common.NewList), Index: 3,
-		ShardId: 0,
-	}
-	shard0Leaving0 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk6"),
-		List:      string(common.LeavingList),
-		ShardId:   0,
-	}
-	shardMetaLeaving1 := &state.ShardValidatorInfo{
-		PublicKey: []byte("pk7"),
-		List:      string(common.LeavingList),
-		Index:     1,
-		ShardId:   core.MetachainShardId,
-	}
-
-	previousConfig := &nodesCoordinator.EpochNodesConfig{
-		EligibleMap: map[uint32][]validator{},
-	}
-
-	validatorInfos :=
-		[]*state.ShardValidatorInfo{
-			shard0Eligible0,
-			shard0Eligible1,
-			shardmetaEligible0,
-			shard0Waiting0,
-			shardmetaWaiting0,
-			shard0New0,
-			shard0Leaving0,
-			shardMetaLeaving1,
-		}
-
-	ihgs.flagWaitingListFix.Unset()
-	newNodesConfig, err := ihgs.computeNodesConfigFromList(previousConfig, validatorInfos)
-	assert.Nil(t, err)
-
-	assert.Equal(t, uint32(1), newNodesConfig.NbShards)
-
-	verifySizes(t, newNodesConfig)
-	verifyLeavingNodesInEligible(t, newNodesConfig)
-
-	// maps have the correct validators inside
-	eligibleListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0Eligible0, shard0Eligible1, shard0Leaving0})
-	assert.Equal(t, eligibleListShardZero, newNodesConfig.EligibleMap[0])
-	eligibleListMeta := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shardmetaEligible0, shardMetaLeaving1})
-	assert.Equal(t, eligibleListMeta, newNodesConfig.EligibleMap[core.MetachainShardId])
-
-	waitingListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0Waiting0})
-	assert.Equal(t, waitingListShardZero, newNodesConfig.WaitingMap[0])
-	waitingListMeta := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shardmetaWaiting0})
-	assert.Equal(t, waitingListMeta, newNodesConfig.WaitingMap[core.MetachainShardId])
-
-	leavingListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0Leaving0})
-	assert.Equal(t, leavingListShardZero, newNodesConfig.LeavingMap[0])
-
-	leavingListMeta := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shardMetaLeaving1})
-	assert.Equal(t, leavingListMeta, newNodesConfig.LeavingMap[core.MetachainShardId])
-
-	newListShardZero := createValidatorList(ihgs,
-		[]*state.ShardValidatorInfo{shard0New0})
-	assert.Equal(t, newListShardZero, newNodesConfig.NewList)
-}
-
-func createValidatorList(ihgs *indexHashedNodesCoordinator, shardValidators []*state.ShardValidatorInfo) []validator {
-	validators := make([]validator, len(shardValidators))
-	for i, v := range shardValidators {
-		shardValidator, _ := nodesCoordinator.NewValidator(
-			v.PublicKey,
-			ihgs.GetChance(v.TempRating),
-			v.Index)
-		validators[i] = shardValidator
-	}
-	sort.Sort(validatorList(validators))
-	return validators
-}
-
-func verifyLeavingNodesInEligible(t *testing.T, newNodesConfig *nodesCoordinator.EpochNodesConfig) {
-	for leavingShardId, leavingValidators := range newNodesConfig.LeavingMap {
-		for _, leavingValidator := range leavingValidators {
-			found, shardId := searchInMap(newNodesConfig.EligibleMap, leavingValidator.PubKey())
-			assert.True(t, found)
-			assert.Equal(t, leavingShardId, shardId)
-		}
-	}
-}
-
-func verifyLeavingNodesInEligibleOrWaiting(t *testing.T, newNodesConfig *nodesCoordinator.EpochNodesConfig) {
-	for leavingShardId, leavingValidators := range newNodesConfig.LeavingMap {
-		for _, leavingValidator := range leavingValidators {
-			found, shardId := searchInMap(newNodesConfig.EligibleMap, leavingValidator.PubKey())
-			if !found {
-				found, shardId = searchInMap(newNodesConfig.WaitingMap, leavingValidator.PubKey())
-			}
-			assert.True(t, found)
-			assert.Equal(t, leavingShardId, shardId)
-		}
-	}
-}
-
-func verifySizes(t *testing.T, newNodesConfig *nodesCoordinator.EpochNodesConfig) {
-	expectedEligibleSize := 2
-	expectedWaitingSize := 2
-	expectedNewSize := 1
-	expectedLeavingSize := 2
-
-	assert.NotNil(t, newNodesConfig)
-	assert.Equal(t, uint32(expectedEligibleSize-1), newNodesConfig.NbShards)
-	assert.Equal(t, expectedEligibleSize, len(newNodesConfig.EligibleMap))
-	assert.Equal(t, expectedWaitingSize, len(newNodesConfig.WaitingMap))
-	assert.Equal(t, expectedNewSize, len(newNodesConfig.NewList))
-	assert.Equal(t, expectedLeavingSize, len(newNodesConfig.LeavingMap))
 }
 
 func TestIndexHashedNodesCoordinator_IsInterfaceNil(t *testing.T) {
