+++ sharding/nodesCoordinator/indexHashedNodesCoordinatorLite_test.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"runtime"
+	"sort"
 	"strconv"
 	"sync"
 	"sync/atomic"
@@ -17,8 +18,10 @@ import (
 	"github.com/ElrondNetwork/elrond-go-core/hashing/sha256"
 	"github.com/ElrondNetwork/elrond-go/common"
 	"github.com/ElrondNetwork/elrond-go/sharding/mock"
+	"github.com/ElrondNetwork/elrond-go/state"
 	"github.com/ElrondNetwork/elrond-go/storage/lrucache"
 	"github.com/ElrondNetwork/elrond-go/testscommon/nodeTypeProviderMock"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -1083,4 +1086,381 @@ func TestIndexHashedNodesCoordinator_IsInterfaceNil(t *testing.T) {
 	ihgs3, err := NewIndexHashedNodesCoordinatorLite(arguments)
 	require.Nil(t, err)
 	require.False(t, check.IfNil(ihgs3))
+}
+
+// ComputeNodesConfigFromList
+
+func TestIndexHashedNodesCoordinator_ComputeNodesConfigFromListNilPreviousNodesConfig(t *testing.T) {
+	t.Parallel()
+
+	arguments := createArguments()
+	pk := []byte("pk")
+	arguments.SelfPublicKey = pk
+	ihncl, _ := NewIndexHashedNodesCoordinatorLite(arguments)
+
+	ihncl.flagWaitingListFix.Unset()
+	validatorInfos := make([]*state.ShardValidatorInfo, 0)
+	newNodesConfig, err := ihncl.ComputeNodesConfigFromList(nil, validatorInfos)
+
+	assert.Nil(t, newNodesConfig)
+	assert.False(t, errors.Is(err, ErrNilPreviousEpochConfig))
+
+	newNodesConfig, err = ihncl.ComputeNodesConfigFromList(nil, nil)
+
+	assert.Nil(t, newNodesConfig)
+	assert.False(t, errors.Is(err, ErrNilPreviousEpochConfig))
+
+	ihncl.flagWaitingListFix.Set()
+	newNodesConfig, err = ihncl.ComputeNodesConfigFromList(nil, validatorInfos)
+
+	assert.Nil(t, newNodesConfig)
+	assert.True(t, errors.Is(err, ErrNilPreviousEpochConfig))
+
+	newNodesConfig, err = ihncl.ComputeNodesConfigFromList(nil, nil)
+
+	assert.Nil(t, newNodesConfig)
+	assert.True(t, errors.Is(err, ErrNilPreviousEpochConfig))
+}
+
+func TestIndexHashedNodesCoordinator_ComputeNodesConfigFromListNoValidators(t *testing.T) {
+	t.Parallel()
+
+	arguments := createArguments()
+	pk := []byte("pk")
+	arguments.SelfPublicKey = pk
+	ihncl, _ := NewIndexHashedNodesCoordinatorLite(arguments)
+
+	validatorInfos := make([]*state.ShardValidatorInfo, 0)
+	newNodesConfig, err := ihncl.ComputeNodesConfigFromList(&EpochNodesConfig{}, validatorInfos)
+
+	assert.Nil(t, newNodesConfig)
+	assert.True(t, errors.Is(err, ErrMapSizeZero))
+
+	newNodesConfig, err = ihncl.ComputeNodesConfigFromList(&EpochNodesConfig{}, nil)
+
+	assert.Nil(t, newNodesConfig)
+	assert.True(t, errors.Is(err, ErrMapSizeZero))
+}
+
+func TestIndexHashedNodesCoordinator_ComputeNodesConfigFromListNilPk(t *testing.T) {
+	t.Parallel()
+
+	arguments := createArguments()
+	pk := []byte("pk")
+	arguments.SelfPublicKey = pk
+	ihncl, _ := NewIndexHashedNodesCoordinatorLite(arguments)
+
+	validatorInfos :=
+		[]*state.ShardValidatorInfo{
+			{
+				PublicKey:  pk,
+				ShardId:    0,
+				List:       "test1",
+				Index:      0,
+				TempRating: 0,
+			},
+			{
+				PublicKey:  nil,
+				ShardId:    0,
+				List:       "test",
+				Index:      0,
+				TempRating: 0,
+			},
+		}
+
+	newNodesConfig, err := ihncl.ComputeNodesConfigFromList(&EpochNodesConfig{}, validatorInfos)
+
+	assert.Nil(t, newNodesConfig)
+	assert.NotNil(t, err)
+	assert.Equal(t, ErrNilPubKey, err)
+}
+
+func TestIndexHashedNodesCoordinator_ComputeNodesConfigFromListValidatorsWithFix(t *testing.T) {
+	t.Parallel()
+
+	arguments := createArguments()
+	pk := []byte("pk")
+	arguments.SelfPublicKey = pk
+	ihgs, _ := NewIndexHashedNodesCoordinatorLite(arguments)
+	ihgs.flagWaitingListFix.Set()
+
+	shard0Eligible0 := &state.ShardValidatorInfo{
+		PublicKey:  []byte("pk0"),
+		List:       string(common.EligibleList),
+		Index:      1,
+		TempRating: 2,
+		ShardId:    0,
+	}
+	shard0Eligible1 := &state.ShardValidatorInfo{
+		PublicKey:  []byte("pk1"),
+		List:       string(common.EligibleList),
+		Index:      2,
+		TempRating: 2,
+		ShardId:    0,
+	}
+	shardmetaEligible0 := &state.ShardValidatorInfo{
+		PublicKey:  []byte("pk2"),
+		ShardId:    core.MetachainShardId,
+		List:       string(common.EligibleList),
+		Index:      1,
+		TempRating: 4,
+	}
+	shard0Waiting0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk3"),
+		List:      string(common.WaitingList),
+		Index:     14,
+		ShardId:   0,
+	}
+	shardmetaWaiting0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk4"),
+		ShardId:   core.MetachainShardId,
+		List:      string(common.WaitingList),
+		Index:     15,
+	}
+	shard0New0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk5"),
+		List:      string(common.NewList), Index: 3,
+		ShardId: 0,
+	}
+	shard0Leaving0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk6"),
+		List:      string(common.LeavingList),
+		ShardId:   0,
+	}
+	shardMetaLeaving1 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk7"),
+		List:      string(common.LeavingList),
+		Index:     1,
+		ShardId:   core.MetachainShardId,
+	}
+
+	validatorInfos :=
+		[]*state.ShardValidatorInfo{
+			shard0Eligible0,
+			shard0Eligible1,
+			shardmetaEligible0,
+			shard0Waiting0,
+			shardmetaWaiting0,
+			shard0New0,
+			shard0Leaving0,
+			shardMetaLeaving1,
+		}
+
+	previousConfig := &EpochNodesConfig{
+		EligibleMap: map[uint32][]Validator{
+			0: {
+				mock.NewValidatorMock(shard0Eligible0.PublicKey, 0, 0),
+				mock.NewValidatorMock(shard0Eligible1.PublicKey, 0, 0),
+				mock.NewValidatorMock(shard0Leaving0.PublicKey, 0, 0),
+			},
+			core.MetachainShardId: {
+				mock.NewValidatorMock(shardmetaEligible0.PublicKey, 0, 0),
+			},
+		},
+		WaitingMap: map[uint32][]Validator{
+			0: {
+				mock.NewValidatorMock(shard0Waiting0.PublicKey, 0, 0),
+			},
+			core.MetachainShardId: {
+				mock.NewValidatorMock(shardmetaWaiting0.PublicKey, 0, 0),
+				mock.NewValidatorMock(shardMetaLeaving1.PublicKey, 0, 0),
+			},
+		},
+	}
+
+	newNodesConfig, err := ihgs.ComputeNodesConfigFromList(previousConfig, validatorInfos)
+	assert.Nil(t, err)
+
+	assert.Equal(t, uint32(1), newNodesConfig.NbShards)
+
+	verifySizes(t, newNodesConfig)
+	verifyLeavingNodesInEligibleOrWaiting(t, newNodesConfig)
+
+	// maps have the correct validators inside
+	eligibleListShardZero := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shard0Eligible0, shard0Eligible1, shard0Leaving0})
+	assert.Equal(t, eligibleListShardZero, newNodesConfig.EligibleMap[0])
+	eligibleListMeta := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shardmetaEligible0})
+	assert.Equal(t, eligibleListMeta, newNodesConfig.EligibleMap[core.MetachainShardId])
+
+	waitingListShardZero := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shard0Waiting0})
+	assert.Equal(t, waitingListShardZero, newNodesConfig.WaitingMap[0])
+	waitingListMeta := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shardmetaWaiting0, shardMetaLeaving1})
+	assert.Equal(t, waitingListMeta, newNodesConfig.WaitingMap[core.MetachainShardId])
+
+	leavingListShardZero := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shard0Leaving0})
+	assert.Equal(t, leavingListShardZero, newNodesConfig.LeavingMap[0])
+
+	leavingListMeta := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shardMetaLeaving1})
+	assert.Equal(t, leavingListMeta, newNodesConfig.LeavingMap[core.MetachainShardId])
+
+	newListShardZero := createValidatorList(ihgs,
+		[]*state.ShardValidatorInfo{shard0New0})
+	assert.Equal(t, newListShardZero, newNodesConfig.NewList)
+}
+
+func TestIndexHashedNodesCoordinator_ComputeNodesConfigFromListValidatorsNoFix(t *testing.T) {
+	t.Parallel()
+
+	arguments := createArguments()
+	pk := []byte("pk")
+	arguments.SelfPublicKey = pk
+	ihncl, _ := NewIndexHashedNodesCoordinatorLite(arguments)
+
+	shard0Eligible0 := &state.ShardValidatorInfo{
+		PublicKey:  []byte("pk0"),
+		List:       string(common.EligibleList),
+		Index:      1,
+		TempRating: 2,
+		ShardId:    0,
+	}
+	shard0Eligible1 := &state.ShardValidatorInfo{
+		PublicKey:  []byte("pk1"),
+		List:       string(common.EligibleList),
+		Index:      2,
+		TempRating: 2,
+		ShardId:    0,
+	}
+	shardmetaEligible0 := &state.ShardValidatorInfo{
+		PublicKey:  []byte("pk2"),
+		ShardId:    core.MetachainShardId,
+		List:       string(common.EligibleList),
+		Index:      1,
+		TempRating: 4,
+	}
+	shard0Waiting0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk3"),
+		List:      string(common.WaitingList),
+		Index:     14,
+		ShardId:   0,
+	}
+	shardmetaWaiting0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk4"),
+		ShardId:   core.MetachainShardId,
+		List:      string(common.WaitingList),
+		Index:     15,
+	}
+	shard0New0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk5"),
+		List:      string(common.NewList), Index: 3,
+		ShardId: 0,
+	}
+	shard0Leaving0 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk6"),
+		List:      string(common.LeavingList),
+		ShardId:   0,
+	}
+	shardMetaLeaving1 := &state.ShardValidatorInfo{
+		PublicKey: []byte("pk7"),
+		List:      string(common.LeavingList),
+		Index:     1,
+		ShardId:   core.MetachainShardId,
+	}
+
+	previousConfig := &EpochNodesConfig{
+		EligibleMap: map[uint32][]Validator{},
+	}
+
+	validatorInfos :=
+		[]*state.ShardValidatorInfo{
+			shard0Eligible0,
+			shard0Eligible1,
+			shardmetaEligible0,
+			shard0Waiting0,
+			shardmetaWaiting0,
+			shard0New0,
+			shard0Leaving0,
+			shardMetaLeaving1,
+		}
+
+	ihncl.flagWaitingListFix.Unset()
+	newNodesConfig, err := ihncl.ComputeNodesConfigFromList(previousConfig, validatorInfos)
+	assert.Nil(t, err)
+
+	assert.Equal(t, uint32(1), newNodesConfig.NbShards)
+
+	verifySizes(t, newNodesConfig)
+	verifyLeavingNodesInEligible(t, newNodesConfig)
+
+	// maps have the correct validators inside
+	eligibleListShardZero := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shard0Eligible0, shard0Eligible1, shard0Leaving0})
+	assert.Equal(t, eligibleListShardZero, newNodesConfig.EligibleMap[0])
+	eligibleListMeta := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shardmetaEligible0, shardMetaLeaving1})
+	assert.Equal(t, eligibleListMeta, newNodesConfig.EligibleMap[core.MetachainShardId])
+
+	waitingListShardZero := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shard0Waiting0})
+	assert.Equal(t, waitingListShardZero, newNodesConfig.WaitingMap[0])
+	waitingListMeta := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shardmetaWaiting0})
+	assert.Equal(t, waitingListMeta, newNodesConfig.WaitingMap[core.MetachainShardId])
+
+	leavingListShardZero := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shard0Leaving0})
+	assert.Equal(t, leavingListShardZero, newNodesConfig.LeavingMap[0])
+
+	leavingListMeta := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shardMetaLeaving1})
+	assert.Equal(t, leavingListMeta, newNodesConfig.LeavingMap[core.MetachainShardId])
+
+	newListShardZero := createValidatorList(ihncl,
+		[]*state.ShardValidatorInfo{shard0New0})
+	assert.Equal(t, newListShardZero, newNodesConfig.NewList)
+}
+
+func createValidatorList(ihncl *IndexHashedNodesCoordinatorLite, shardValidators []*state.ShardValidatorInfo) []Validator {
+	validators := make([]Validator, len(shardValidators))
+	for i, v := range shardValidators {
+		shardValidator, _ := NewValidator(
+			v.PublicKey,
+			ihncl.GetChance(v.TempRating),
+			v.Index)
+		validators[i] = shardValidator
+	}
+	sort.Sort(ValidatorList(validators))
+	return validators
+}
+
+func verifyLeavingNodesInEligibleOrWaiting(t *testing.T, newNodesConfig *EpochNodesConfig) {
+	for leavingShardId, leavingValidators := range newNodesConfig.LeavingMap {
+		for _, leavingValidator := range leavingValidators {
+			found, shardId := searchInMap(newNodesConfig.EligibleMap, leavingValidator.PubKey())
+			if !found {
+				found, shardId = searchInMap(newNodesConfig.WaitingMap, leavingValidator.PubKey())
+			}
+			assert.True(t, found)
+			assert.Equal(t, leavingShardId, shardId)
+		}
+	}
+}
+
+func verifyLeavingNodesInEligible(t *testing.T, newNodesConfig *EpochNodesConfig) {
+	for leavingShardId, leavingValidators := range newNodesConfig.LeavingMap {
+		for _, leavingValidator := range leavingValidators {
+			found, shardId := searchInMap(newNodesConfig.EligibleMap, leavingValidator.PubKey())
+			assert.True(t, found)
+			assert.Equal(t, leavingShardId, shardId)
+		}
+	}
+}
+
+func verifySizes(t *testing.T, newNodesConfig *EpochNodesConfig) {
+	expectedEligibleSize := 2
+	expectedWaitingSize := 2
+	expectedNewSize := 1
+	expectedLeavingSize := 2
+
+	assert.NotNil(t, newNodesConfig)
+	assert.Equal(t, uint32(expectedEligibleSize-1), newNodesConfig.NbShards)
+	assert.Equal(t, expectedEligibleSize, len(newNodesConfig.EligibleMap))
+	assert.Equal(t, expectedWaitingSize, len(newNodesConfig.WaitingMap))
+	assert.Equal(t, expectedNewSize, len(newNodesConfig.NewList))
+	assert.Equal(t, expectedLeavingSize, len(newNodesConfig.LeavingMap))
 }
